+++ CHANGELOG.md
@@ -70,6 +70,7 @@ Developers may need to change their code
 
 * Plugin context check before and after all callbacks.
   * Check blocked signals and signal handlers
+  * Check termios settings
   * Any changes to context are logged at loglevel WARNING
 * Added set/get pointer API to clixon_data:
    * clicon_ptr_get(), clicon_ptr_set(),